code-samples/calendar/python/get_calendar_availability/pyproject.toml
[tool.poetry]
name = "get-calendar-availability"
version = "0.1.0"
description = "Get calendar availability using Nylas Python SDK"
authors = ["wiseaidev <oss@wiseai.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nylas = { version = "<7", allow-prereleases = true}
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.10.0"
pytest = "^7.4.2"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get_calendar_availability = "get_calendar_availability.__main__:main"

[virtualenvs]
in-project = true
create = true